fix(plugin): recursive optimisation of pattern matches in attributes/properties (#41)
Also fixed the spawning of `singleton` operations in pattern matches.

Previously, reactivity of solid-js would be broken when pattern matching on a property directly. This is because the generation of the expression `(matchValue = props.___, matchValue === ____ ? ____ : ___)` et al would not be caught by the solid-js library/compiler.

For this reason, we remove this CE optimisation when pattern matching on property getters only.

This behaviour can be disabled using the SkipCEOptimisation component flag.

This constitutes a 'fix' more than a breaking change.