feat!: Storybook Generation (#40)
* storybook base implementation

* alpha version storybook

* doc extensions, member val string enum matching, et al

* known bug: render and arg variant pairs don't mesh well. Better to just stick to one or the other.

Made it so key value pairs for children arg will have the value expr transformed by the main plugin

* fix fantomas break

* Remove jetbrains language injection attributes to determine if they are preventing
tailwind injection by the tailwind language service

* bump version

* fix(plugin): recursive optimisation of pattern matches in attributes/properties (#41)

Also fixed the spawning of `singleton` operations in pattern matches.

Previously, reactivity of solid-js would be broken when pattern matching on a property directly. This is because the generation of the expression `(matchValue = props.___, matchValue === ____ ? ____ : ___)` et al would not be caught by the solid-js library/compiler.

For this reason, we remove this CE optimisation when pattern matching on property getters only.

This behaviour can be disabled using the SkipCEOptimisation component flag.

This constitutes a 'fix' more than a breaking change.

* fix(plugin): ArrayAlloc and ArrayFrom expressions are transformed

* better expression collection abstraction

* add(storybook): `decorator` custom operation for defining decorators

* add(storybook): `decorator` custom operation for defining decorators

* feat: storybook support